rename table_dump related modules for consistency
