+++ src/parser/mrt/messages/mod.rs
@@ -1,15 +1,25 @@
-use crate::models::{AsnLength, Bgp4MpEnum, Bgp4MpType, MrtMessage};
+use crate::models::{
+    AsnLength, Bgp4MpEnum, Bgp4MpType, MrtMessage, TableDumpV2Message, TableDumpV2Type,
+};
 use bytes::Bytes;
 
 pub(crate) mod bgp4mp;
-pub(crate) mod table_dump_message;
-pub(crate) mod table_dump_v2_message;
+pub(crate) mod table_dump;
+pub(crate) mod table_dump_v2;
 
 impl MrtMessage {
     pub fn encode(&self, sub_type: u16) -> Bytes {
         let msg_bytes: Bytes = match self {
             MrtMessage::TableDumpMessage(m) => m.encode(),
-            MrtMessage::TableDumpV2Message(_m) => {
+            MrtMessage::TableDumpV2Message(m) => {
+                let v2_type: TableDumpV2Type = TableDumpV2Type::try_from(sub_type).unwrap();
+                match m {
+                    TableDumpV2Message::PeerIndexTable(p) => {}
+                    TableDumpV2Message::RibAfi(r) => {}
+                    TableDumpV2Message::RibGeneric(_) => {
+                        todo!("RibGeneric message is not supported yet");
+                    }
+                }
                 todo!("TableDumpV2 message is not supported yet");
             }
             MrtMessage::Bgp4Mp(m) => {
